Try undo formFields logic
