+++ src/components/UserLogin/Login.jsx
@@ -1,45 +1,79 @@
-import { useState } from "react";
+import axios from "axios";
+import { useState, useContext } from "react";
+import { useNavigate } from "react-router-dom";
+import { AuthContext } from "../../context/auth.context";
 import { loginFields } from "../../constants/formFields";
 import Input from "./Input";
 import FormAction from "./FormAction";
 import FormExtra from "./FormExtra";
+
+// USER AUTH
+const API_URL = "http://localhost:5005";
 
 const fields = loginFields;
 let fieldsState = {};
 fields.forEach((field) => (fieldsState[field.id] = ""));
 
 export default function Login() {
+  // const [email, setEmail] = useState("");
+  // const [password, setPassword] = useState("");
   const [loginState, setLoginState] = useState(fieldsState);
-
   const handleChange = (e) => {
     setLoginState({ ...loginState, [e.target.id]: e.target.value });
   };
 
-  const handleSubmit = (e) => {
+  console.log(loginState);
+
+  const [errorMessage, setErrorMessage] = useState(undefined);
+
+  const navigate = useNavigate();
+
+  const { storeToken, verifyUser } = useContext(AuthContext);
+
+  const handleLoginSubmit = (e) => {
     e.preventDefault();
-    authenticateUser();
+
+    const loginBody = { email, password };
+
+    axios
+      .post(`${API_URL}/login`, loginBody)
+      .then((res) => {
+        console.log("JWT token", res.data.authToken);
+
+        storeToken(res.data.authToken);
+
+        verifyUser();
+        navigate("/");
+      })
+      .catch((err) => {
+        console.log(err);
+        setErrorMessage(err.response.data.errorMessage);
+      });
   };
 
   return (
-    <form className="mt-8 space-y-6 onSubmit={handleSubmit}">
-      <div className="-space-y-px">
-        {fields.map((field) => (
-          <Input
-            key={field.id}
-            handleChange={handleChange}
-            value={loginState[field.id]}
-            labelText={field.labelText}
-            labelFor={field.labelFor}
-            id={field.id}
-            name={field.name}
-            type={field.type}
-            isRequired={field.isRequired}
-            placeholder={field.placeholder}
-          />
-        ))}
-      </div>
-      <FormExtra />
-      <FormAction handleSubmit={handleSubmit} text="Login" />
-    </form>
+    <div>
+      <form className="mt-8 space-y-6 onSubmit={handleLoginSubmit}">
+        <div className="-space-y-px">
+          {fields.map((field) => (
+            <Input
+              key={field.id}
+              handleChange={handleChange}
+              value={loginState[field.id]}
+              labelText={field.labelText}
+              labelFor={field.labelFor}
+              id={field.id}
+              name={field.name}
+              type={field.type}
+              isRequired={field.isRequired}
+              placeholder={field.placeholder}
+            />
+          ))}
+        </div>
+        <FormExtra />
+        <FormAction handleLoginSubmit={handleLoginSubmit} text="Login" />
+      </form>
+      {errorMessage && <p className="error-message">{errorMessage}</p>}
+    </div>
   );
 }
